@@ -78,6 +78,8 @@ class Phake
      */
     const CLIENT_DEFAULT = 'DEFAULT';
     const CLIENT_PHPUNIT = 'PHPUNIT';
+    const CLIENT_PHPUNIT6 = 'PHPUNIT6';
+    const CLIENT_PHPUNIT7 = 'PHPUNIT7';
 
     /**
      * Returns a new mock object based on the given class name.
@@ -447,7 +449,7 @@ class Phake
      */
     public static function equalTo($value)
     {
-        return new Phake_Matchers_EqualsMatcher($value, new \SebastianBergmann\Comparator\Factory());
+        return new Phake_Matchers_EqualsMatcher($value, \SebastianBergmann\Comparator\Factory::getInstance());
     }
 
     /**
@@ -567,7 +569,13 @@ class Phake
     public static function getClient()
     {
         if (!isset(self::$client)) {
-            if (class_exists('PHPUnit_Framework_TestCase')) {
+            if (class_exists('PHPUnit\Framework\TestCase')) {
+                if (7 <= \PHPUnit\Runner\Version::id()) {
+                    return self::$client = new Phake_Client_PHPUnit7();
+                } else {
+                    return self::$client = new Phake_Client_PHPUnit6();
+                }
+            } elseif (class_exists('PHPUnit_Framework_TestCase')) {
                 return self::$client = new Phake_Client_PHPUnit();
             }
             return self::$client = new Phake_Client_Default();
@@ -589,7 +597,12 @@ class Phake
             self::$client = $client;
         } elseif ($client == self::CLIENT_PHPUNIT) {
             self::$client = new Phake_Client_PHPUnit();
-        } else {
+        } elseif ($client == self::CLIENT_PHPUNIT6) {
+            self::$client = new Phake_Client_PHPUnit6();
+        } elseif ($client == self::CLIENT_PHPUNIT7) {
+            self::$client = new Phake_Client_PHPUnit7();
+        }
+        else {
             self::$client = new Phake_Client_Default();
         }
     }
