@@ -2,26 +2,26 @@
 
 /*
  * Phake - Mocking Framework
- * 
+ *
  * Copyright (c) 2010, Mike Lively <mike.lively@sellingsource.com>
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  *  *  Redistributions of source code must retain the above copyright
  *     notice, this list of conditions and the following disclaimer.
- * 
+ *
  *  *  Redistributions in binary form must reproduce the above copyright
  *     notice, this list of conditions and the following disclaimer in
  *     the documentation and/or other materials provided with the
  *     distribution.
- * 
+ *
  *  *  Neither the name of Mike Lively nor the names of his
  *     contributors may be used to endorse or promote products derived
  *     from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
@@ -34,7 +34,7 @@
  * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
- * 
+ *
  * @category   Testing
  * @package    Phake
  * @author     Mike Lively <m@digitalsandwich.com>
@@ -43,7 +43,11 @@
  * @link       http://www.digitalsandwich.com/
  */
 
-class Phake_ClassGenerator_InvocationHandler_FrozenObjectCheckTest extends PHPUnit_Framework_TestCase
+use PHPUnit\Framework\ExpectationFailedException;
+
+use PHPUnit\Framework\TestCase;
+
+class Phake_ClassGenerator_InvocationHandler_FrozenObjectCheckTest extends TestCase
 {
     /**
      * @var Phake_ClassGenerator_InvocationHandler_FrozenObjectCheck
@@ -58,13 +62,9 @@ class Phake_ClassGenerator_InvocationHandler_FrozenObjectCheckTest extends PHPUn
 
     public function setUp()
     {
+        Phake::setClient(Phake::CLIENT_PHPUNIT7);
         Phake::initAnnotations($this);
         $this->handler    = new Phake_ClassGenerator_InvocationHandler_FrozenObjectCheck($this->mockInfo);
-    }
-
-    protected function tearDown()
-    {
-        Phake::setClient(Phake::CLIENT_DEFAULT);
     }
 
     public function testImplementIInvocationHandler()
@@ -74,7 +74,8 @@ class Phake_ClassGenerator_InvocationHandler_FrozenObjectCheckTest extends PHPUn
 
     public function testReturnsWithNoIssuesIfObjectIsNotFrozen()
     {
-        $mock = $this->getMock('Phake_IMock');
+        $mock = $this->getMockBuilder('Phake_IMock')
+                    ->getMock();
         Phake::when($this->mockInfo)->isObjectFrozen()->thenReturn(false);
 
         try {
@@ -83,26 +84,28 @@ class Phake_ClassGenerator_InvocationHandler_FrozenObjectCheckTest extends PHPUn
         } catch (Exception $e) {
             $this->fail('There should not have been an exception:' . $e->getMessage());
         }
+        Phake::verify($this->mockInfo)->isObjectFrozen();
     }
 
     public function testThrowsWhenObjectIsFrozen()
     {
-        $mock = $this->getMock('Phake_IMock');
+        $mock = $this->getMockBuilder('Phake_IMock')
+                    ->getMock();
         Phake::when($this->mockInfo)->isObjectFrozen()->thenReturn(true);
 
-        $this->setExpectedException('Phake_Exception_VerificationException', 'This object has been frozen.');
+        $this->expectException('Phake_Exception_VerificationException', 'This object has been frozen.');
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         $ref = array();
         $this->handler->invoke($mock, 'foo', array(), $ref);
     }
 
     public function testThrowsWhenObjectIsFrozenWithPHPUnit()
     {
-        Phake::setClient(Phake::CLIENT_PHPUNIT);
-
-        $mock = $this->getMock('Phake_IMock');
+        $mock = $this->getMockBuilder('Phake_IMock')
+                    ->getMock();
         Phake::when($this->mockInfo)->isObjectFrozen()->thenReturn(true);
 
-        $this->setExpectedException('PHPUnit_Framework_ExpectationFailedException', 'This object has been frozen.');
+        $this->expectException(ExpectationFailedException::class, 'This object has been frozen.');
         $ref = array();
         $this->handler->invoke($mock, 'foo', array(), $ref);
     }
