@@ -2,26 +2,26 @@
 
 /*
  * Phake - Mocking Framework
- * 
+ *
  * Copyright (c) 2010, Mike Lively <mike.lively@sellingsource.com>
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  *  *  Redistributions of source code must retain the above copyright
  *     notice, this list of conditions and the following disclaimer.
- * 
+ *
  *  *  Redistributions in binary form must reproduce the above copyright
  *     notice, this list of conditions and the following disclaimer in
  *     the documentation and/or other materials provided with the
  *     distribution.
- * 
+ *
  *  *  Neither the name of Mike Lively nor the names of his
  *     contributors may be used to endorse or promote products derived
  *     from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
@@ -34,7 +34,7 @@
  * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
- * 
+ *
  * @category   Testing
  * @package    Phake
  * @author     Mike Lively <m@digitalsandwich.com>
@@ -43,7 +43,9 @@
  * @link       http://www.digitalsandwich.com/
  */
 
-class Phake_ClassGenerator_InvocationHandler_MagicCallRecorderTest extends PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+class Phake_ClassGenerator_InvocationHandler_MagicCallRecorderTest extends TestCase
 {
     /**
      * @var Phake_ClassGenerator_InvocationHandler_MagicCallRecorder
@@ -68,7 +70,8 @@ class Phake_ClassGenerator_InvocationHandler_MagicCallRecorderTest extends PHPUn
 
     public function testMagicCallIsRecorded()
     {
-        $mock = $this->getMock('Phake_IMock');
+        $mock = $this->getMockBuilder('Phake_IMock')
+                    ->getMock();
 
         $ref = array('foo', array());
         $this->handler->invoke($mock, '__call', array('foo', array()), $ref);
@@ -80,7 +83,8 @@ class Phake_ClassGenerator_InvocationHandler_MagicCallRecorderTest extends PHPUn
 
     public function testStaticMagicCallIsRecorded()
     {
-        $mock = $this->getMock('Phake_IMock');
+        $mock = $this->getMockBuilder('Phake_IMock')
+                    ->getMock();
         $mockClass = get_class($mock);
 
         $ref = array('foo', array());
@@ -93,7 +97,8 @@ class Phake_ClassGenerator_InvocationHandler_MagicCallRecorderTest extends PHPUn
 
     public function testNonMagicCallDoesNothing()
     {
-        $mock = $this->getMock('Phake_IMock');
+        $mock = $this->getMockBuilder('Phake_IMock')
+                    ->getMock();
 
         $ref = array();
         $this->handler->invoke($mock, 'foo', array(), $ref);
@@ -101,4 +106,3 @@ class Phake_ClassGenerator_InvocationHandler_MagicCallRecorderTest extends PHPUn
         Phake::verifyNoInteraction($this->callRecorder);
     }
 }
-
