@@ -2,26 +2,26 @@
 
 /*
  * Phake - Mocking Framework
- * 
+ *
  * Copyright (c) 2010, Mike Lively <mike.lively@sellingsource.com>
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  *  *  Redistributions of source code must retain the above copyright
  *     notice, this list of conditions and the following disclaimer.
- * 
+ *
  *  *  Redistributions in binary form must reproduce the above copyright
  *     notice, this list of conditions and the following disclaimer in
  *     the documentation and/or other materials provided with the
  *     distribution.
- * 
+ *
  *  *  Neither the name of Mike Lively nor the names of his
  *     contributors may be used to endorse or promote products derived
  *     from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
@@ -34,7 +34,7 @@
  * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
- * 
+ *
  * @category   Testing
  * @package    Phake
  * @author     Mike Lively <m@digitalsandwich.com>
@@ -43,12 +43,21 @@
  * @link       http://www.digitalsandwich.com/
  */
 
-class Phake_Client_PHPUnitTest extends PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\ExpectationFailedException;
+use PHPUnit\Framework\Assert;
+use PHPUnit\Runner\Version;
+
+class Phake_Client_PHPUnitTest extends TestCase
 {
     private $client;
 
     public function setUp()
     {
+        if (version_compare(Version::id(), '6.0.0') >= 0) {
+            $this->markTestSkipped('The tested class is not compatible with current version of PHPUnit.');
+        }
+
         $this->client = new Phake_Client_PHPUnit();
     }
 
@@ -59,6 +68,9 @@ class Phake_Client_PHPUnitTest extends PHPUnit_Framework_TestCase
 
     public function testProcessVerifierResultReturnsCallsOnTrue()
     {
+        if (version_compare('6.0.0', Version::id()) != 1) {
+            $this->markTestSkipped('The tested class is not compatible with current version of PHPUnit.');
+        }
         $result = new Phake_CallRecorder_VerifierResult(true, array('call1'));
 
         $this->assertEquals(array('call1'), $this->client->processVerifierResult($result));
@@ -68,7 +80,7 @@ class Phake_Client_PHPUnitTest extends PHPUnit_Framework_TestCase
     {
         $result = new Phake_CallRecorder_VerifierResult(false, array(), 'failure message');
 
-        $this->setExpectedException('PHPUnit_Framework_ExpectationFailedException', 'failure message');
+        $this->expectException(ExpectationFailedException::class, 'failure message');
         $this->client->processVerifierResult($result);
     }
 
@@ -76,9 +88,9 @@ class Phake_Client_PHPUnitTest extends PHPUnit_Framework_TestCase
     {
         $result = new Phake_CallRecorder_VerifierResult(true, array('call1'));
 
-        $assertionCount = PHPUnit_Framework_Assert::getCount();
+        $assertionCount = Assert::getCount();
         $this->client->processVerifierResult($result);
-        $newAssertionCount = PHPUnit_Framework_Assert::getCount();
+        $newAssertionCount = Assert::getCount();
 
         $this->assertGreaterThan($assertionCount, $newAssertionCount);
     }
@@ -88,9 +100,9 @@ class Phake_Client_PHPUnitTest extends PHPUnit_Framework_TestCase
      */
     public function testProcessObjectFreeze()
     {
-        $assertionCount = PHPUnit_Framework_Assert::getCount();
+        $assertionCount = Assert::getCount();
         $this->client->processObjectFreeze();
-        $newAssertionCount = PHPUnit_Framework_Assert::getCount();
+        $newAssertionCount = Assert::getCount();
 
         $this->assertGreaterThan($assertionCount, $newAssertionCount);
     }
