@@ -1,26 +1,26 @@
 <?php
-/* 
+/*
  * Phake - Mocking Framework
- * 
+ *
  * Copyright (c) 2010-2012, Mike Lively <m@digitalsandwich.com>
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  *  *  Redistributions of source code must retain the above copyright
  *     notice, this list of conditions and the following disclaimer.
- * 
+ *
  *  *  Redistributions in binary form must reproduce the above copyright
  *     notice, this list of conditions and the following disclaimer in
  *     the documentation and/or other materials provided with the
  *     distribution.
- * 
+ *
  *  *  Neither the name of Mike Lively nor the names of his
  *     contributors may be used to endorse or promote products derived
  *     from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
@@ -33,7 +33,7 @@
  * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
- * 
+ *
  * @category   Testing
  * @package    Phake
  * @author     Mike Lively <m@digitalsandwich.com>
@@ -42,7 +42,10 @@
  * @link       http://www.digitalsandwich.com/
  */
 
-class Phake_Matchers_FactoryTest extends PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\Constraint\Constraint;
+
+class Phake_Matchers_FactoryTest extends TestCase
 {
     /**
      * @var Phake_Matchers_Factory
@@ -75,7 +78,7 @@ class Phake_Matchers_FactoryTest extends PHPUnit_Framework_TestCase
      */
     public function testPassThroughMatcher()
     {
-        $matcher = $this->getMock('Phake_Matchers_IChainableArgumentMatcher');
+        $matcher = $this->getMockBuilder('Phake_Matchers_IChainableArgumentMatcher')->getMock();
 
         $retMatcher = $this->factory->createMatcher($matcher);
 
@@ -85,9 +88,9 @@ class Phake_Matchers_FactoryTest extends PHPUnit_Framework_TestCase
     /**
      * Tests creating a phpunit adapter matcher
      */
-    public function testPHPUnitConstraint()
+    public function testPHPUnit6Constraint()
     {
-        $matcher = $this->getMock('PHPUnit_Framework_Constraint');
+        $matcher = $this->getMockBuilder(Constraint::class)->getMock();
         $matcher->expects($this->once())
             ->method('evaluate')
             ->with($this->equalTo('foo'))
@@ -104,7 +107,7 @@ class Phake_Matchers_FactoryTest extends PHPUnit_Framework_TestCase
      */
     public function testHamcrestMatcher()
     {
-        $matcher = $this->getMock('Hamcrest\Matcher');
+        $matcher = $this->getMockBuilder('Hamcrest\Matcher')->getMock();
         $matcher->expects($this->once())
             ->method('matches')
             ->with($this->equalTo('foo'))
@@ -158,5 +161,3 @@ class Phake_Matchers_FactoryTest extends PHPUnit_Framework_TestCase
         $this->assertNull($this->factory->createMatcherChain(array()));
     }
 }
-
-
