@@ -42,7 +42,9 @@
  * @link       http://www.digitalsandwich.com/
  */
 
-class Phake_Proxies_CallStubberProxyTest extends PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+class Phake_Proxies_CallStubberProxyTest extends TestCase
 {
     /**
      * @var Phake_Proxies_CallStubberProxy
@@ -65,7 +67,7 @@ class Phake_Proxies_CallStubberProxyTest extends PHPUnit_Framework_TestCase
     public function setUp()
     {
         $this->matcher1 = Phake::mock('Phake_Matchers_IChainableArgumentMatcher');
-        $this->obj      = $this->getMock('Phake_IMock');
+        $this->obj      = $this->getMockBuilder('Phake_IMock')->getMock();
         PhakeTestUtil::setStubMapper($this->obj, Phake::mock('Phake_Stubber_StubMapper'));
         $this->proxy    = new Phake_Proxies_CallStubberProxy($this->matcher1, false);
     }
