@@ -2,26 +2,26 @@
 
 /*
  * Phake - Mocking Framework
- * 
+ *
  * Copyright (c) 2010, Mike Lively <mike.lively@sellingsource.com>
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  *  *  Redistributions of source code must retain the above copyright
  *     notice, this list of conditions and the following disclaimer.
- * 
+ *
  *  *  Redistributions in binary form must reproduce the above copyright
  *     notice, this list of conditions and the following disclaimer in
  *     the documentation and/or other materials provided with the
  *     distribution.
- * 
+ *
  *  *  Neither the name of Mike Lively nor the names of his
  *     contributors may be used to endorse or promote products derived
  *     from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
@@ -34,7 +34,7 @@
  * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
- * 
+ *
  * @category   Testing
  * @package    Phake
  * @author     Mike Lively <m@digitalsandwich.com>
@@ -64,7 +64,7 @@ class Phake_Annotation_MockInitializer
         $properties = $reader->getPropertiesWithAnnotation('Mock');
 
         foreach ($properties as $property) {
-            $annotations = $reader->getPropertyAnnotations($property);
+            $annotations = $reader->readReflectionAnnotation($property);
 
             if ($annotations['Mock'] !== true) {
                 $mockedClass = $annotations['Mock'];
@@ -75,24 +75,29 @@ class Phake_Annotation_MockInitializer
             if (isset($parser)) {
                 // Ignore it if the class start with a backslash
                 if (substr($mockedClass, 0, 1) !== '\\') {
-                    $useStatements = $parser->parseClass($reflectionClass);
+                    $useStatements = $parser->parseClass($property->getDeclaringClass());
                     $key           = strtolower($mockedClass);
 
                     if (array_key_exists($key, $useStatements)) {
                         $mockedClass = $useStatements[$key];
+                    } elseif ($reflectionClass->getNamespaceName() && $this->definedUnderTestNamespace($mockedClass, $reflectionClass->getNamespaceName())) {
+                        $mockedClass = $reflectionClass->getNamespaceName() . '\\' . $mockedClass;
                     }
                 }
             }
 
-            $reflProp = new ReflectionProperty(get_class($object), $property);
-
-            $reflProp->setAccessible(true);
-            $reflProp->setValue($object, Phake::mock($mockedClass));
+            $property->setAccessible(true);
+            $property->setValue($object, Phake::mock($mockedClass));
         }
     }
 
     protected function useDoctrineParser()
     {
         return version_compare(PHP_VERSION, "5.3.3", ">=") && class_exists('Doctrine\Common\Annotations\PhpParser');
+    }
+
+    protected function definedUnderTestNamespace($mockedClass, $testNamespace)
+    {
+        return class_exists($testNamespace . '\\' . $mockedClass);
     }
 }
