@@ -2,26 +2,26 @@
 
 /*
  * Phake - Mocking Framework
- * 
+ *
  * Copyright (c) 2010, Mike Lively <mike.lively@sellingsource.com>
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  *  *  Redistributions of source code must retain the above copyright
  *     notice, this list of conditions and the following disclaimer.
- * 
+ *
  *  *  Redistributions in binary form must reproduce the above copyright
  *     notice, this list of conditions and the following disclaimer in
  *     the documentation and/or other materials provided with the
  *     distribution.
- * 
+ *
  *  *  Neither the name of Mike Lively nor the names of his
  *     contributors may be used to endorse or promote products derived
  *     from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
@@ -34,7 +34,7 @@
  * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
- * 
+ *
  * @category   Testing
  * @package    Phake
  * @author     Mike Lively <m@digitalsandwich.com>
@@ -82,7 +82,7 @@ class Phake_Annotation_Reader
      *
      * @param string $annotation
      *
-     * @return array
+     * @return ReflectionProperty[]
      */
     public function getPropertiesWithAnnotation($annotation)
     {
@@ -91,7 +91,7 @@ class Phake_Annotation_Reader
             $annotations = $this->getPropertyAnnotations($property->getName());
 
             if (array_key_exists($annotation, $annotations)) {
-                $properties[] = $property->getName();
+                $properties[] = $property;
             }
         }
         return $properties;
@@ -106,7 +106,7 @@ class Phake_Annotation_Reader
      *
      * @return array
      */
-    private function readReflectionAnnotation($reflVar)
+    public function readReflectionAnnotation($reflVar)
     {
         $comment = $reflVar->getDocComment();
 
