@@ -52,7 +52,7 @@ class Phake_Matchers_Factory
 
     public function __construct()
     {
-        $this->comparatorFactory = new \SebastianBergmann\Comparator\Factory();
+        $this->comparatorFactory =  \SebastianBergmann\Comparator\Factory::getInstance();
     }
 
     /**
@@ -75,7 +75,11 @@ class Phake_Matchers_Factory
         } elseif ($argument instanceof PHPUnit_Framework_Constraint) {
             $return = new Phake_Matchers_PHPUnitConstraintAdapter($argument);
         } elseif ($argument instanceof \PHPUnit\Framework\Constraint\Constraint) {
-            $return = new Phake_Matchers_PHPUnit6ConstraintAdapter($argument);
+            if (7 <= \PHPUnit\Runner\Version::id()) {
+                $return = new Phake_Matchers_PHPUnit7ConstraintAdapter($argument);
+            } else {
+                $return = new Phake_Matchers_PHPUnit6ConstraintAdapter($argument);
+            }
         } elseif ($argument instanceof Hamcrest\Matcher) {
             $return = new Phake_Matchers_HamcrestMatcherAdapter($argument);
         } elseif ($argument instanceof Phake_Matchers_IArgumentMatcher) {
