@@ -1,26 +1,26 @@
 <?php
-/* 
+/*
  * Phake - Mocking Framework
- * 
+ *
  * Copyright (c) 2010-2012, Mike Lively <m@digitalsandwich.com>
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  *  *  Redistributions of source code must retain the above copyright
  *     notice, this list of conditions and the following disclaimer.
- * 
+ *
  *  *  Redistributions in binary form must reproduce the above copyright
  *     notice, this list of conditions and the following disclaimer in
  *     the documentation and/or other materials provided with the
  *     distribution.
- * 
+ *
  *  *  Neither the name of Mike Lively nor the names of his
  *     contributors may be used to endorse or promote products derived
  *     from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
@@ -33,7 +33,7 @@
  * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
- * 
+ *
  * @category   Testing
  * @package    Phake
  * @author     Mike Lively <m@digitalsandwich.com>
@@ -42,12 +42,14 @@
  * @link       http://www.digitalsandwich.com/
  */
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Description of VerifierTest
  *
  * @author Mike Lively <m@digitalsandwich.com>
  */
-class Phake_CallRecorder_VerifierTest extends PHPUnit_Framework_TestCase
+class Phake_CallRecorder_VerifierTest extends TestCase
 {
     /**
      * @var Phake_CallRecorder_Recorder
@@ -147,8 +149,8 @@ class Phake_CallRecorder_VerifierTest extends PHPUnit_Framework_TestCase
      */
     public function testVerifierDoesNotFindCallWithUnmatchedArguments()
     {
-        $matcher1 = new Phake_Matchers_EqualsMatcher('test', new \SebastianBergmann\Comparator\Factory());
-        $matcher2 = new Phake_Matchers_EqualsMatcher('test', new \SebastianBergmann\Comparator\Factory());
+        $matcher1 = new Phake_Matchers_EqualsMatcher('test', \SebastianBergmann\Comparator\Factory::getInstance());
+        $matcher2 = new Phake_Matchers_EqualsMatcher('test', \SebastianBergmann\Comparator\Factory::getInstance());
         $matcher1->setNextMatcher($matcher2);
         $expectation = new Phake_CallRecorder_CallExpectation(
             $this->obj,
@@ -224,8 +226,11 @@ class Phake_CallRecorder_VerifierTest extends PHPUnit_Framework_TestCase
     public function testVerifierBeingCalledWithMixedCallRecorder()
     {
         $recorder = new Phake_CallRecorder_Recorder();
-        $obj1     = $this->getMock('Phake_IMock');
-        $obj2     = $this->getMock('Phake_IMock');
+
+        $obj1     = $this->getMockBuilder('Phake_IMock')
+                        ->getMock();
+        $obj2     = $this->getMockBuilder('Phake_IMock')
+                        ->getMock();
 
         $expectation = new Phake_CallRecorder_CallExpectation(
             $obj1,
@@ -332,7 +337,7 @@ Other Invocations:
         $expectation = new Phake_CallRecorder_CallExpectation(
             $this->obj,
             'foo',
-            new Phake_Matchers_EqualsMatcher('test', new \SebastianBergmann\Comparator\Factory()),
+            new Phake_Matchers_EqualsMatcher('test', \SebastianBergmann\Comparator\Factory::getInstance()),
             $this->verifierMode
         );
 
