@@ -1,26 +1,26 @@
 <?php
-/* 
+/*
  * Phake - Mocking Framework
- * 
+ *
  * Copyright (c) 2010-2012, Mike Lively <m@digitalsandwich.com>
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  *  *  Redistributions of source code must retain the above copyright
  *     notice, this list of conditions and the following disclaimer.
- * 
+ *
  *  *  Redistributions in binary form must reproduce the above copyright
  *     notice, this list of conditions and the following disclaimer in
  *     the documentation and/or other materials provided with the
  *     distribution.
- * 
+ *
  *  *  Neither the name of Mike Lively nor the names of his
  *     contributors may be used to endorse or promote products derived
  *     from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
@@ -33,7 +33,7 @@
  * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
- * 
+ *
  * @category   Testing
  * @package    Phake
  * @author     Mike Lively <m@digitalsandwich.com>
@@ -42,10 +42,12 @@
  * @link       http://www.digitalsandwich.com/
  */
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Tests the functionality of the equals matcher
  */
-class Phake_Matchers_EqualsMatcherTest extends PHPUnit_Framework_TestCase
+class Phake_Matchers_EqualsMatcherTest extends TestCase
 {
     /**
      * @var Phake_Matchers_EqualsMatcher
@@ -57,7 +59,7 @@ class Phake_Matchers_EqualsMatcherTest extends PHPUnit_Framework_TestCase
      */
     public function setUp()
     {
-        $this->matcher = new Phake_Matchers_EqualsMatcher('foo', new \SebastianBergmann\Comparator\Factory());
+        $this->matcher = new Phake_Matchers_EqualsMatcher('foo', \SebastianBergmann\Comparator\Factory::getInstance());
     }
 
     /**
@@ -75,7 +77,7 @@ class Phake_Matchers_EqualsMatcherTest extends PHPUnit_Framework_TestCase
     public function testBadMatches()
     {
         $value = array('test');
-        $this->setExpectedException('Exception');
+        $this->expectException('Exception');
         $this->matcher->doArgumentsMatch($value);
     }
 
@@ -91,7 +93,7 @@ class Phake_Matchers_EqualsMatcherTest extends PHPUnit_Framework_TestCase
      */
     public function testToStringOnNonStringableObject()
     {
-        $this->matcher = new Phake_Matchers_EqualsMatcher(new stdClass, new \SebastianBergmann\Comparator\Factory());
+        $this->matcher = new Phake_Matchers_EqualsMatcher(new stdClass, \SebastianBergmann\Comparator\Factory::getInstance());
 
         $this->assertEquals('equal to <object:stdClass>', $this->matcher->__toString());
     }
@@ -104,7 +106,7 @@ class Phake_Matchers_EqualsMatcherTest extends PHPUnit_Framework_TestCase
         $a             = new stdClass;
         $a->b          = new stdClass;
         $a->b->a       = $a;
-        $this->matcher = new Phake_Matchers_EqualsMatcher($a, new \SebastianBergmann\Comparator\Factory());
+        $this->matcher = new Phake_Matchers_EqualsMatcher($a, \SebastianBergmann\Comparator\Factory::getInstance());
 
         $c       = new stdClass();
         $c->b    = new stdClass();
@@ -117,28 +119,28 @@ class Phake_Matchers_EqualsMatcherTest extends PHPUnit_Framework_TestCase
 
     public function testDifferentClassObjects()
     {
-        $this->matcher = new Phake_Matchers_EqualsMatcher(new PhakeTest_A(), new \SebastianBergmann\Comparator\Factory());
+        $this->matcher = new Phake_Matchers_EqualsMatcher(new PhakeTest_A(), \SebastianBergmann\Comparator\Factory::getInstance());
 
         $value = array(new PhakeTest_B());
-        $this->setExpectedException('Exception');
+        $this->expectException('Exception');
         $this->matcher->doArgumentsMatch($value);
     }
 
     public function testArraysWithDifferentCounts()
     {
-        $this->matcher = new Phake_Matchers_EqualsMatcher(array(1), new \SebastianBergmann\Comparator\Factory());
+        $this->matcher = new Phake_Matchers_EqualsMatcher(array(1), \SebastianBergmann\Comparator\Factory::getInstance());
 
         $test = array(array(1, 2));
-        $this->setExpectedException('Phake_Exception_MethodMatcherException');
+        $this->expectException('Phake_Exception_MethodMatcherException');
         $this->matcher->doArgumentsMatch($test);
     }
 
     public function testArraysWithDifferentKeys()
     {
-        $this->matcher = new Phake_Matchers_EqualsMatcher(array('one' => 1), new \SebastianBergmann\Comparator\Factory());
+        $this->matcher = new Phake_Matchers_EqualsMatcher(array('one' => 1), \SebastianBergmann\Comparator\Factory::getInstance());
 
         $test = array(array('two' => 1));
-        $this->setExpectedException('Exception');
+        $this->expectException('Exception');
         $this->matcher->doArgumentsMatch($test);
     }
 }
