@@ -1,26 +1,26 @@
 <?php
-/* 
+/*
  * Phake - Mocking Framework
- * 
+ *
  * Copyright (c) 2010-2012, Mike Lively <m@digitalsandwich.com>
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  *  *  Redistributions of source code must retain the above copyright
  *     notice, this list of conditions and the following disclaimer.
- * 
+ *
  *  *  Redistributions in binary form must reproduce the above copyright
  *     notice, this list of conditions and the following disclaimer in
  *     the documentation and/or other materials provided with the
  *     distribution.
- * 
+ *
  *  *  Neither the name of Mike Lively nor the names of his
  *     contributors may be used to endorse or promote products derived
  *     from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
@@ -33,7 +33,7 @@
  * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
- * 
+ *
  * @category   Testing
  * @package    Phake
  * @author     Mike Lively <m@digitalsandwich.com>
@@ -42,12 +42,14 @@
  * @link       http://www.digitalsandwich.com/
  */
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Description of AnswerBinderProxyTest
  *
  * @author Mike Lively <m@digitalsandwich.com>
  */
-class Phake_Proxies_AnswerBinderProxyTest extends PHPUnit_Framework_TestCase
+class Phake_Proxies_AnswerBinderProxyTest extends TestCase
 {
     /**
      * @var Phake_Proxies_AnswerBinderProxy
@@ -64,7 +66,9 @@ class Phake_Proxies_AnswerBinderProxyTest extends PHPUnit_Framework_TestCase
      */
     public function setUp()
     {
-        $this->binder = $this->getMock('Phake_Stubber_AnswerBinder', array(), array(), '', false);
+        $this->binder = $this->getMockBuilder('Phake_Stubber_AnswerBinder')
+                            ->disableOriginalConstructor()
+                            ->getMock();
         $this->proxy  = new Phake_Proxies_AnswerBinderProxy($this->binder);
     }
 
@@ -97,7 +101,9 @@ class Phake_Proxies_AnswerBinderProxyTest extends PHPUnit_Framework_TestCase
      */
     public function testThenReturnCallback()
     {
-        $func = create_function('$arg1', 'return $arg1;');
+        $func = function($arg1) {
+            return $arg1;
+        };
 
         $this->binder->expects($this->once())
             ->method('bindAnswer')
@@ -117,7 +123,7 @@ class Phake_Proxies_AnswerBinderProxyTest extends PHPUnit_Framework_TestCase
      */
     public function testThenReturnCallbackThrowsExceptionForUncallableLambda()
     {
-        $this->setExpectedException('InvalidArgumentException');
+        $this->expectException('InvalidArgumentException');
 
         $func = 'some_unknown_function';
         $this->proxy->thenReturnCallback($func);
