@@ -1,26 +1,26 @@
 <?php
-/* 
+/*
  * Phake - Mocking Framework
- * 
+ *
  * Copyright (c) 2010, Mike Lively <mike.lively@sellingsource.com>
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  *  *  Redistributions of source code must retain the above copyright
  *     notice, this list of conditions and the following disclaimer.
- * 
+ *
  *  *  Redistributions in binary form must reproduce the above copyright
  *     notice, this list of conditions and the following disclaimer in
  *     the documentation and/or other materials provided with the
  *     distribution.
- * 
+ *
  *  *  Neither the name of Mike Lively nor the names of his
  *     contributors may be used to endorse or promote products derived
  *     from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
@@ -33,7 +33,7 @@
  * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
- * 
+ *
  * @category   Testing
  * @package    Phake
  * @author     Mike Lively <m@digitalsandwich.com>
@@ -42,8 +42,9 @@
  * @link       http://www.digitalsandwich.com/
  */
 
+use PHPUnit\Framework\TestCase;
 
-class Phake_Proxies_AnswerCollectionProxyTest extends PHPUnit_Framework_TestCase
+class Phake_Proxies_AnswerCollectionProxyTest extends TestCase
 {
     /**
      * @var Phake_Proxies_AnswerCollectionProxyTest
@@ -90,7 +91,9 @@ class Phake_Proxies_AnswerCollectionProxyTest extends PHPUnit_Framework_TestCase
      */
     public function testThenGetReturnByLambda()
     {
-        $func = create_function('$arg1', 'return $arg1;');
+        $func = function($arg1) {
+            return $arg1;
+        };
 
         $this->assertSame($this->proxy, $this->proxy->thenGetReturnByLambda($func));
 
@@ -107,7 +110,7 @@ class Phake_Proxies_AnswerCollectionProxyTest extends PHPUnit_Framework_TestCase
      */
     public function testThenGetReturnByLambdaThrowsExceptionForUncallableLambda()
     {
-        $this->setExpectedException('InvalidArgumentException');
+        $this->expectException('InvalidArgumentException');
 
         $func = 'some_unknown_function';
         $this->proxy->thenGetReturnByLambda($func);
