@@ -1,26 +1,26 @@
 <?php
-/* 
+/*
  * Phake - Mocking Framework
- * 
+ *
  * Copyright (c) 2010-2012, Mike Lively <m@digitalsandwich.com>
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  *  *  Redistributions of source code must retain the above copyright
  *     notice, this list of conditions and the following disclaimer.
- * 
+ *
  *  *  Redistributions in binary form must reproduce the above copyright
  *     notice, this list of conditions and the following disclaimer in
  *     the documentation and/or other materials provided with the
  *     distribution.
- * 
+ *
  *  *  Neither the name of Mike Lively nor the names of his
  *     contributors may be used to endorse or promote products derived
  *     from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
@@ -33,7 +33,7 @@
  * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
- * 
+ *
  * @category   Testing
  * @package    Phake
  * @author     Mike Lively <m@digitalsandwich.com>
@@ -42,12 +42,14 @@
  * @link       http://www.digitalsandwich.com/
  */
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Description of VerifierProxyTest
  *
  * @author Mike Lively <m@digitalsandwich.com>
  */
-class Phake_Proxies_VerifierProxyTest extends PHPUnit_Framework_TestCase
+class Phake_Proxies_VerifierProxyTest extends TestCase
 {
     /**
      * @var Phake_CallRecorder_Verifier
@@ -80,7 +82,7 @@ class Phake_Proxies_VerifierProxyTest extends PHPUnit_Framework_TestCase
         );
 
         $this->proxy = new Phake_Proxies_VerifierProxy($this->verifier, new Phake_Matchers_Factory(), $this->mode, $this->client);
-        $obj         = $this->getMock('Phake_IMock');
+        $obj         = $this->getMockBuilder('Phake_IMock')->getMock();
         Phake::when($this->verifier)->getObject()->thenReturn($obj);
         Phake::when($this->mode)->__toString()->thenReturn('exactly 1 times');
         Phake::when($this->client)->processVerifierResult($this->anything())->thenReturn($this->matchedCalls);
@@ -134,7 +136,7 @@ class Phake_Proxies_VerifierProxyTest extends PHPUnit_Framework_TestCase
      */
     public function testProxyTransformsNonMatchersToEqualsMatcher()
     {
-        $argumentMatcher = new Phake_Matchers_EqualsMatcher('test', new \SebastianBergmann\Comparator\Factory());
+        $argumentMatcher = new Phake_Matchers_EqualsMatcher('test', \SebastianBergmann\Comparator\Factory::getInstance());
         Phake::when($this->verifier)->verifyCall(Phake::anyParameters())->thenReturn(
             new Phake_CallRecorder_VerifierResult(true, array(Phake::mock('Phake_CallRecorder_CallInfo')))
         );
@@ -159,7 +161,7 @@ class Phake_Proxies_VerifierProxyTest extends PHPUnit_Framework_TestCase
      */
     public function testMagicGetWithInvalidData($invalidData, $exceptionContains)
     {
-        $this->setExpectedException('InvalidArgumentException', $exceptionContains);
+        $this->expectException('InvalidArgumentException', $exceptionContains);
         $this->proxy->__get($invalidData);
     }
 
