@@ -1,26 +1,26 @@
 <?php
-/* 
+/*
  * Phake - Mocking Framework
- * 
+ *
  * Copyright (c) 2010-2012, Mike Lively <m@digitalsandwich.com>
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  *  *  Redistributions of source code must retain the above copyright
  *     notice, this list of conditions and the following disclaimer.
- * 
+ *
  *  *  Redistributions in binary form must reproduce the above copyright
  *     notice, this list of conditions and the following disclaimer in
  *     the documentation and/or other materials provided with the
  *     distribution.
- * 
+ *
  *  *  Neither the name of Mike Lively nor the names of his
  *     contributors may be used to endorse or promote products derived
  *     from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
@@ -33,7 +33,7 @@
  * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
- * 
+ *
  * @category   Testing
  * @package    Phake
  * @author     Mike Lively <m@digitalsandwich.com>
@@ -42,12 +42,14 @@
  * @link       http://www.digitalsandwich.com/
  */
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Tests the Answer Factory
  *
  * @author Mike Lively <m@digitalsandwich.com>
  */
-class Phake_Stubber_AnswerBinderTest extends PHPUnit_Framework_TestCase
+class Phake_Stubber_AnswerBinderTest extends TestCase
 {
     /**
      * @var Phake_Stubber_AnswerBinder
@@ -70,7 +72,9 @@ class Phake_Stubber_AnswerBinderTest extends PHPUnit_Framework_TestCase
     public function setUp()
     {
         Phake::initAnnotations($this);
-        $this->matcher    = $this->getMock('Phake_Matchers_MethodMatcher', array(), array(), '', false);
+        $this->matcher    = $this->getMockBuilder('Phake_Matchers_MethodMatcher')
+                                ->disableOriginalConstructor()
+                                ->getMock();
         $this->stubMapper = Phake::mock('Phake_Stubber_StubMapper');
 
         $this->binder = new Phake_Stubber_AnswerBinder($this->matcher, $this->stubMapper);
@@ -78,7 +82,7 @@ class Phake_Stubber_AnswerBinderTest extends PHPUnit_Framework_TestCase
 
     public function testBindAnswer()
     {
-        $answer = $this->getMock('Phake_Stubber_IAnswer');
+        $answer = $this->getMockBuilder('Phake_Stubber_IAnswer')->getMock();
 
         $this->binder->bindAnswer($answer);
 
@@ -87,7 +91,7 @@ class Phake_Stubber_AnswerBinderTest extends PHPUnit_Framework_TestCase
 
     public function testBindAnswerReturnsAnswerCollectionBinder()
     {
-        $answer = $this->getMock('Phake_Stubber_IAnswer');
+        $answer = $this->getMockBuilder('Phake_Stubber_IAnswer')->getMock();
 
         $newBinder = $this->binder->bindAnswer($answer);
 
