@@ -42,6 +42,9 @@
  * @link       http://www.digitalsandwich.com/
  */
 
+use PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\ExpectationFailedException;
+
 /**
  * Tests the behavior of the Phake class.
  *
@@ -49,17 +52,17 @@
  *
  * @author Mike Lively <m@digitalsandwich.com>
  */
-class PhakeTest extends PHPUnit_Framework_TestCase
+class PhakeTest extends TestCase
 {
     public function setUp()
     {
-        Phake::setClient(Phake::CLIENT_DEFAULT);
+        Phake::setClient(Phake::CLIENT_PHPUNIT7);
     }
 
     protected function tearDown()
     {
         Phake::resetStaticInfo();
-        Phake::setClient(Phake::CLIENT_DEFAULT);
+        Phake::setClient(Phake::CLIENT_PHPUNIT7);
     }
 
     /**
@@ -90,6 +93,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
      */
     public function testSimpleVerifyThrowsExceptionOnFail()
     {
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         $mock = Phake::mock('PhakeTest_MockedClass');
 
         Phake::verify($mock)->foo();
@@ -257,6 +261,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
      */
     public function testVerifyCallWithEqualsMatcherFails()
     {
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         $mock = Phake::mock('PhakeTest_MockedClass');
 
         $mock->fooWithArgument('test');
@@ -283,6 +288,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
      */
     public function testVerifyCallWithDefaultMatcherFails()
     {
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         $mock = Phake::mock('PhakeTest_MockedClass');
 
         $mock->fooWithArgument('test');
@@ -309,6 +315,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
      */
     public function testVerifyCallWithPHPUnitMatcherFails()
     {
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         $mock = Phake::mock('PhakeTest_MockedClass');
 
         $mock->fooWithArgument('test');
@@ -335,6 +342,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
      */
     public function testVerifyCallWithHamcrestMatcherFails()
     {
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         $mock = Phake::mock('PhakeTest_MockedClass');
 
         $mock->fooWithArgument('test');
@@ -407,7 +415,8 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         Phake::reset($mock);
 
-        $this->setExpectedException('Phake_Exception_VerificationException');
+        $this->expectException('Phake_Exception_VerificationException');
+        Phake::setClient(Phake::CLIENT_DEFAULT);
 
         Phake::verify($mock)->foo();
     }
@@ -441,7 +450,8 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         Phake::resetStatic($mock);
 
-        $this->setExpectedException('Phake_Exception_VerificationException');
+        $this->expectException('Phake_Exception_VerificationException');
+        Phake::setClient(Phake::CLIENT_DEFAULT);
 
         Phake::verifyStatic($mock)->staticMethod();
     }
@@ -695,7 +705,8 @@ class PhakeTest extends PHPUnit_Framework_TestCase
         $mock->callInnerFunc();
         $mock->fooWithReturnValue();
 
-        $this->setExpectedException('Phake_Exception_VerificationException');
+        $this->expectException('Phake_Exception_VerificationException');
+        Phake::setClient(Phake::CLIENT_DEFAULT);
 
         Phake::inOrder(
             Phake::verify($mock)->foo(),
@@ -742,7 +753,8 @@ class PhakeTest extends PHPUnit_Framework_TestCase
         $mock2->fooWithReturnValue();
         $mock2->callInnerFunc();
 
-        $this->setExpectedException('Phake_Exception_VerificationException');
+        $this->expectException('Phake_Exception_VerificationException');
+        Phake::setClient(Phake::CLIENT_DEFAULT);
 
         Phake::inOrder(
             Phake::verify($mock2)->fooWithReturnValue(),
@@ -778,7 +790,8 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         Phake::verifyNoFurtherInteraction($mock);
 
-        $this->setExpectedException('Phake_Exception_VerificationException');
+        $this->expectException('Phake_Exception_VerificationException');
+        Phake::setClient(Phake::CLIENT_DEFAULT);
 
         $mock->foo();
     }
@@ -791,7 +804,8 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         Phake::verifyNoFurtherInteraction($mock);
 
-        $this->setExpectedException('Phake_Exception_VerificationException');
+        $this->expectException('Phake_Exception_VerificationException');
+        Phake::setClient(Phake::CLIENT_DEFAULT);
 
         $mock::staticMethod();
     }
@@ -809,7 +823,8 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         Phake::verifyNoFurtherInteraction($mock1, $mock2);
 
-        $this->setExpectedException('Phake_Exception_VerificationException');
+        $this->expectException('Phake_Exception_VerificationException');
+        Phake::setClient(Phake::CLIENT_DEFAULT);
 
         $mock2->foo();
     }
@@ -825,7 +840,8 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         $mock->foo();
 
-        $this->setExpectedException('Phake_Exception_VerificationException');
+        $this->expectException('Phake_Exception_VerificationException');
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         Phake::verifyNoInteraction($mock);
     }
 
@@ -840,7 +856,8 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         $mock::staticMethod();
 
-        $this->setExpectedException('Phake_Exception_VerificationException');
+        $this->expectException('Phake_Exception_VerificationException');
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         Phake::verifyNoInteraction($mock);
     }
 
@@ -856,7 +873,8 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         $mock2->foo();
 
-        $this->setExpectedException('Phake_Exception_VerificationException');
+        $this->expectException('Phake_Exception_VerificationException');
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         Phake::verifyNoInteraction($mock1, $mock2);
     }
 
@@ -900,7 +918,8 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         $mock->fooWithArgument('FOO');
 
-        $this->setExpectedException('Phake_Exception_VerificationException');
+        $this->expectException('Phake_Exception_VerificationException');
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         Phake::verify($mock)->fooWithArgument(Phake::capture($toArgument)->when('BAR'));
     }
 
@@ -1056,6 +1075,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
      */
     public function testVerifyTimesMismatch()
     {
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         $mock = Phake::mock('PhakeTest_MockedClass');
 
         $mock->foo();
@@ -1095,6 +1115,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
      */
     public function testVerifyAtLeastMismatch()
     {
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         $mock = Phake::mock('PhakeTest_MockedClass');
 
         Phake::verify($mock, Phake::atLeast(1))->foo();
@@ -1115,6 +1136,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
      */
     public function testNeverMismatch()
     {
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         $mock = Phake::mock('PhakeTest_MockedClass');
         $mock->foo();
         Phake::verify($mock, Phake::never())->foo();
@@ -1145,6 +1167,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
      */
     public function testAtMostOver()
     {
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         $mock = Phake::mock('PhakeTest_MockedClass');
         $mock->foo();
         $mock->foo();
@@ -1195,7 +1218,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         $mock->fooWithLotsOfParameters(3, 2, 1);
 
-        $this->setExpectedException('InvalidArgumentException', 'Other matchers cannot be passed with any '
+        $this->expectException('InvalidArgumentException', 'Other matchers cannot be passed with any '
             . 'parameters. It will not work the way you think it works');
         Phake::verify($mock)->fooWithLotsOfParameters(Phake::anyParameters(), 1);
     }
@@ -1206,7 +1229,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         $mock->fooWithLotsOfParameters(3, 2, 1);
 
-        $this->setExpectedException('InvalidArgumentException', 'Other matchers cannot be passed with any '
+        $this->expectException('InvalidArgumentException', 'Other matchers cannot be passed with any '
             . 'parameters. It will not work the way you think it works');
         Phake::verify($mock)->fooWithLotsOfParameters(3, Phake::anyParameters());
     }
@@ -1231,7 +1254,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         $mock->fooWithLotsOfParameters(3, 2, 1);
 
-        $this->setExpectedException('InvalidArgumentException', 'Other matchers cannot be checked after you ignore remaining parameters.');
+        $this->expectException('InvalidArgumentException', 'Other matchers cannot be checked after you ignore remaining parameters.');
         Phake::verify($mock)->fooWithLotsOfParameters(Phake::ignoreRemaining(), 1);
     }
 
@@ -1252,10 +1275,11 @@ class PhakeTest extends PHPUnit_Framework_TestCase
     {
         $mock = Phake::mock('PhakeTest_MockedClass');
 
-        $this->setExpectedException(
+        $this->expectException(
             'Phake_Exception_VerificationException',
             'Expected PhakeTest_MockedClass->foo() to be called exactly <1> times, actually called <0> times. In fact, there are no interactions with this mock.'
         );
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         Phake::verify($mock)->foo();
     }
 
@@ -1265,7 +1289,8 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         $mock->foo('test');
 
-        $this->setExpectedException(
+        Phake::setClient(Phake::CLIENT_DEFAULT);
+        $this->expectException(
             'Phake_Exception_VerificationException',
             'Expected PhakeTest_MockedClass->foo() to be called exactly <1> times, actually called <0> times.' . "\n"
                 . "Other Invocations:\n"
@@ -1363,16 +1388,16 @@ class PhakeTest extends PHPUnit_Framework_TestCase
     {
         $original_client = Phake::getClient();
 
-        Phake::setClient(Phake::CLIENT_PHPUNIT);
+        Phake::setClient(Phake::CLIENT_PHPUNIT7);
 
-        $this->assertInstanceOf('Phake_Client_PHPUnit', Phake::getClient());
+        $this->assertInstanceOf('Phake_Client_PHPUnit7', Phake::getClient());
 
         Phake::setClient($original_client);
     }
 
     public function testVerifyNoFurtherInteractionPassesStrict()
     {
-        Phake::setClient(Phake::CLIENT_PHPUNIT);
+        Phake::setClient(Phake::CLIENT_PHPUNIT7);
         $mock = Phake::mock('stdClass');
 
         $assertionCount = self::getCount();
@@ -1384,7 +1409,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
     public function testVerifyNoInteractionPassesStrict()
     {
-        Phake::setClient(Phake::CLIENT_PHPUNIT);
+        Phake::setClient(Phake::CLIENT_PHPUNIT7);
         $mock = Phake::mock('stdClass');
 
         $assertionCount = self::getCount();
@@ -1473,7 +1498,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
     public function testCallOrderInObjectFailsWithPHPUnit()
     {
-        Phake::setClient(Phake::CLIENT_PHPUNIT);
+        Phake::setClient(Phake::CLIENT_PHPUNIT7);
 
         $mock = Phake::mock('PhakeTest_MockedClass');
 
@@ -1481,7 +1506,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
         $mock->callInnerFunc();
         $mock->fooWithReturnValue();
 
-        $this->setExpectedException('PHPUnit_Framework_ExpectationFailedException');
+        $this->expectException(ExpectationFailedException::class);
 
         Phake::inOrder(
             Phake::verify($mock)->foo(),
@@ -1583,7 +1608,7 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
 
         eval('
-            $this->setExpectedException("InvalidArgumentException");
+            $this->expectException("InvalidArgumentException");
             $mock = Phake::mock("PhakeTest_MockedClass");
             try
             {
@@ -1604,7 +1629,8 @@ class PhakeTest extends PHPUnit_Framework_TestCase
         $mock->foo('b');
 
         Phake::verify($mock)->foo('a');
-        $this->setExpectedException('Phake_Exception_VerificationException');
+        $this->expectException('Phake_Exception_VerificationException');
+        Phake::setClient(Phake::CLIENT_DEFAULT);
         Phake::verifyNoOtherInteractions($mock);
     }
 
@@ -1708,5 +1734,18 @@ class PhakeTest extends PHPUnit_Framework_TestCase
 
         $val = $obj->askSomething();
         $this->assertEquals('Asked', $val);
+    }
+
+    /**
+     * For #239
+     */
+    public function testChainingDoNothing()
+    {
+        $mock = Phake::mock(PhakeTest_MockedClass::class);
+
+        Phake::when($mock)->foo->thenReturn(42)->thenDoNothing();
+
+        $this->assertEquals(42, $mock->foo());
+        $this->assertNull($mock->foo());
     }
 }
